Make the fixture available from conftest
